outline new environments modal structure
