+++ app/components/shared-new/Modal/Header.tsx
@@ -1,0 +1,26 @@
+import { Box } from "grommet";
+import { copy } from "../../../theme/copy";
+import Close from "../icons/Close";
+import Button from "../AppButton";
+import Text from "../Text";
+
+type Props = {
+  closeModal: () => void;
+  label: string;
+};
+
+export default function Header({ closeModal, label }: Props): JSX.Element {
+  return (
+    <Box align="center" direction="row" flex={false} justify="between">
+      <Text color="gray9" size="componentHeader">
+        {label}
+      </Text>
+      <Button
+        IconComponent={Close}
+        a11yTitle={copy.close}
+        onClick={closeModal}
+        type="ghost"
+      />
+    </Box>
+  );
+}
