+++ app/lib/types.ts
@@ -165,7 +165,6 @@ export type GroupTests = {
 export type Modal =
   | "apiKeys"
   | "deployment"
-  | "envVariables"
   | "environments"
   | "deleteGroup"
   | "deleteTest"
